refactor: shift makeSignatureValidator to common for both mobile and email validator
