+++ src/app/utils/field-validation/validators/common.ts
@@ -1,7 +1,10 @@
 import { left, right } from 'fp-ts/lib/Either'
 
 import { ProcessedSingleAnswerResponse } from 'src/app/modules/submission/submission.types'
+import { IEmailFieldSchema, IMobileFieldSchema } from 'src/types/field'
 import { ResponseValidator } from 'src/types/field/utils/validation'
+
+import formsgSdk from '../../../../config/formsg-sdk'
 
 export const notEmptySingleAnswerResponse: ResponseValidator<ProcessedSingleAnswerResponse> = (
   response,
@@ -11,4 +14,37 @@ export const notEmptySingleAnswerResponse: ResponseValidator<ProcessedSingleAnsw
       'CommonValidator.notEmptySingleAnswerResponse:\tanswer is an empty string',
     )
   return right(response)
+}
+
+type SignatureValidatorConstructor = (
+  formField: IEmailFieldSchema | IMobileFieldSchema,
+) => ResponseValidator<ProcessedSingleAnswerResponse>
+
+export const makeSignatureValidator: SignatureValidatorConstructor = (
+  formField,
+) => (response) => {
+  const { isVerifiable, _id } = formField
+  if (!isVerifiable) {
+    return right(response) // no validation occurred
+  }
+  const { signature, answer } = response
+  if (!signature) {
+    return left(
+      `CommonValidator.makeSignatureValidator:\t answer does not have valid signature`,
+    )
+  }
+  const isSigned =
+    formsgSdk.verification.authenticate &&
+    formsgSdk.verification.authenticate({
+      signatureString: signature,
+      submissionCreatedAt: Date.now(),
+      fieldId: _id,
+      answer,
+    })
+
+  return isSigned
+    ? right(response)
+    : left(
+        `CommonValidator.makeSignatureValidator:\t answer does not have valid signature`,
+      )
 }
